+++ src/Service/UnitConverter/UnitConverterService.php
@@ -10,9 +10,9 @@ final class UnitConverterService implements UnitConverterInterface
     /**
      * @param float $amount
      * @param UnitType $unit
-     * @return int
+     * @return float
      */
-    public function toGrams(float $amount, UnitType $unit): int
+    public function toGrams(float $amount, UnitType $unit): float
     {
         return match ($unit) {
             UnitType::Kilogram => (int)round($amount * 1000.0),
@@ -21,11 +21,11 @@ final class UnitConverterService implements UnitConverterInterface
     }
 
     /**
-     * @param int $grams
+     * @param float $grams
      * @param UnitType $targetUnit
      * @return float
      */
-    public function fromGrams(int $grams, UnitType $targetUnit): float
+    public function fromGrams(float $grams, UnitType $targetUnit): float
     {
         return match ($targetUnit) {
             UnitType::Kilogram => $grams / 1000.0,
